新增权限申请模块 新增worker执行后消息通知模块 todo web端收到worker消息后，通过web socket推送浏览器 todo 权限申请、审批事件通知，通过web socket推送浏览器

+++ web-server/src/main/java/com/nxin/framework/controller/auth/ApplyController.java
@@ -9,6 +9,8 @@ import com.nxin.framework.entity.auth.Apply;
 import com.nxin.framework.entity.auth.Privilege;
 import com.nxin.framework.entity.auth.User;
 import com.nxin.framework.enums.Constant;
+import com.nxin.framework.exception.ExistedException;
+import com.nxin.framework.exception.RecordsNotMatchException;
 import com.nxin.framework.service.auth.ApplyService;
 import com.nxin.framework.service.auth.PrivilegeService;
 import com.nxin.framework.service.auth.ResourceService;
@@ -19,6 +21,7 @@ import com.nxin.framework.vo.auth.ApplyVo;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -45,6 +48,8 @@ public class ApplyController {
     @Autowired
     private UserService userService;
 
+    private static final String AUDIT_STATUS_APPLY = "0";
+
     private static final BeanConverter<ApplyVo, Apply> applyConverter = new ApplyConverter();
 
     @GetMapping("/apply/{id}")
@@ -61,7 +66,7 @@ public class ApplyController {
     public ResponseEntity<PageVo<ApplyVo>> page(@RequestBody ApplyDto applyDto) {
         IPage<Apply> applyIPage = applyService.search(applyDto.getSearchAudit(), applyDto.getCreator(), applyDto.getPageNo(), applyDto.getPageSize());
         List<ApplyVo> applyVos = applyConverter.convert(applyIPage.getRecords());
-        if (applyIPage.getSize() > 0) {
+        if (!applyIPage.getRecords().isEmpty()) {
             LambdaQueryWrapper<Privilege> privilegeLambdaQueryWrapper = new LambdaQueryWrapper<>();
             privilegeLambdaQueryWrapper.in(Privilege::getId, applyIPage.getRecords().stream().map(Apply::getPrivilegeId).collect(Collectors.toList()));
             privilegeLambdaQueryWrapper.eq(Privilege::getStatus, Constant.ACTIVE);
@@ -79,20 +84,46 @@ public class ApplyController {
     public ResponseEntity save(@RequestBody ApplyDto applyDto) {
         Apply apply = new Apply();
         BeanUtils.copyProperties(applyDto, apply);
-        applyService.save(apply);
-        return ResponseEntity.ok().build();
+        apply.setAuditStatus(AUDIT_STATUS_APPLY);
+        try {
+            applyService.save(apply);
+            return ResponseEntity.ok().build();
+        } catch (RecordsNotMatchException e) {
+            return ResponseEntity.status(Constant.EXCEPTION_NOT_FOUNT).build();
+        } catch (ExistedException e) {
+            return ResponseEntity.status(Constant.EXCEPTION_DUPLICATED).body(e.getMessage());
+        }
     }
 
     @PostMapping("/audit")
     public ResponseEntity<ApplyVo> audit(@RequestBody ApplyDto applyDto) {
-        Apply apply = applyService.one(applyDto.getId());
-        if (apply != null) {
+        if (!CollectionUtils.isEmpty(applyDto.getApplyIdList())) {
             User loginUser = userService.one(LoginUtils.getUsername());
-            Privilege privilege = privilegeService.findByPrivilegeIdAndUserId(apply.getPrivilegeId(), loginUser.getId());
-            if (privilege != null) {
-                BeanUtils.copyProperties(applyDto, apply);
-                applyService.audit(apply);
-                return ResponseEntity.ok().build();
+            LambdaQueryWrapper<Apply> applyLambdaQueryWrapper = new LambdaQueryWrapper<>();
+            applyLambdaQueryWrapper.eq(Apply::getStatus, Constant.ACTIVE);
+            applyLambdaQueryWrapper.eq(Apply::getAuditStatus, AUDIT_STATUS_APPLY);
+            applyLambdaQueryWrapper.in(Apply::getId, applyDto.getApplyIdList());
+            List<Apply> applies = applyService.list(applyLambdaQueryWrapper);
+            if (!applies.isEmpty()) {
+                List<User> creatorList = userService.findByEmail(applies.stream().map(Apply::getCreator).collect(Collectors.toList()));
+                LambdaQueryWrapper<Privilege> privilegeLambdaQueryWrapper = new LambdaQueryWrapper<>();
+                privilegeLambdaQueryWrapper.eq(Privilege::getStatus, Constant.ACTIVE);
+                privilegeLambdaQueryWrapper.in(Privilege::getId, applies.stream().map(Apply::getPrivilegeId).collect(Collectors.toList()));
+                List<Privilege> privilegeList = privilegeService.list(privilegeLambdaQueryWrapper);
+                if (!privilegeList.isEmpty()) {
+                    List<Privilege> privileges = privilegeService.findByPrivilegeIdListAndUserId(privilegeList.stream().map(Privilege::getId).collect(Collectors.toList()), loginUser.getId());
+                    if (!privileges.isEmpty()) {
+                        Map<Long, Privilege> privilegeMap = privileges.stream().collect(Collectors.toMap(Privilege::getId, v -> v));
+                        applies.removeIf(item -> !privilegeMap.containsKey(item.getPrivilegeId()));
+                    }
+                }
+                if (!applies.isEmpty()) {
+                    for (Apply apply : applies) {
+                        apply.setAuditStatus(applyDto.getAuditStatus());
+                    }
+                    applyService.audit(applies, creatorList);
+                    return ResponseEntity.ok().build();
+                }
             }
             return ResponseEntity.status(Constant.EXCEPTION_UNAUTHORIZED).build();
         }
@@ -102,7 +133,7 @@ public class ApplyController {
     @DeleteMapping("/apply/{id}")
     public ResponseEntity<ApplyVo> delete(@PathVariable Long id) {
         Apply persisted = applyService.one(id);
-        if (persisted != null) {
+        if (persisted != null && AUDIT_STATUS_APPLY.equals(persisted.getAuditStatus())) {
             if (persisted.getCreator().equals(LoginUtils.getUsername())) {
                 applyService.delete(persisted);
                 return ResponseEntity.ok().build();
